+++ rmp-serde/src/lib.rs
@@ -188,6 +188,87 @@ impl<'de> Deserialize<'de> for Raw {
     }
 }
 
+/// Helper that allows to decode strings no matter whether they contain valid or invalid UTF-8.
+#[derive(Clone, Copy, Debug, PartialEq)]
+pub struct RawRef<'a> {
+    s: Result<&'a str, (&'a [u8], Utf8Error)>,
+}
+
+impl<'a> RawRef<'a> {
+    /// Returns `true` if the raw is valid UTF-8.
+    pub fn is_str(&self) -> bool {
+        self.s.is_ok()
+    }
+
+    /// Returns `true` if the raw contains invalid UTF-8 sequence.
+    pub fn is_err(&self) -> bool {
+        self.s.is_err()
+    }
+
+    /// Returns the string reference if the raw is valid UTF-8, or else `None`.
+    pub fn as_str(&self) -> Option<&str> {
+        match self.s {
+            Ok(ref s) => Some(s),
+            Err(..) => None,
+        }
+    }
+
+    /// Returns the underlying `Utf8Error` if the raw contains invalid UTF-8 sequence, or
+    /// else `None`.
+    pub fn as_err(&self) -> Option<&Utf8Error> {
+        match self.s {
+            Ok(..) => None,
+            Err((_, ref err)) => Some(&err),
+        }
+    }
+
+    /// Returns a byte slice of this raw's contents.
+    pub fn as_bytes(&self) -> &[u8] {
+        match self.s {
+            Ok(ref s) => s.as_bytes(),
+            Err(ref err) => &err.0[..],
+        }
+    }
+}
+
+struct RawRefVisitor;
+
+impl<'de> de::Visitor<'de> for RawRefVisitor {
+    type Value = RawRef<'de>;
+
+    fn expecting(&self, fmt: &mut Formatter) -> Result<(), fmt::Error> {
+        "string or bytes".fmt(fmt)
+    }
+
+    #[inline]
+    fn visit_borrowed_str<E>(self, v: &'de str) -> Result<Self::Value, E>
+        where E: de::Error
+    {
+        Ok(RawRef { s: Ok(v) })
+    }
+
+    #[inline]
+    fn visit_borrowed_bytes<E>(self, v: &'de [u8]) -> Result<Self::Value, E>
+        where E: de::Error
+    {
+        let s = match str::from_utf8(v) {
+            Ok(s) => Ok(s),
+            Err(err) => Err((v, err)),
+        };
+
+        Ok(RawRef { s: s })
+    }
+}
+
+impl<'de> Deserialize<'de> for RawRef<'de> {
+    #[inline]
+    fn deserialize<D>(de: D) -> Result<Self, D::Error>
+        where D: de::Deserializer<'de>
+    {
+        de.deserialize_any(RawRefVisitor)
+    }
+}
+
 /// Serializes a value to a byte vector.
 pub fn to_vec<T>(value: &T) -> Result<Vec<u8>, encode::Error>
     where T: serde::Serialize